Interpreter memory improvements

+++ interpreter/src/main/java/org/qbicc/interpreter/impl/VmClassLoaderImpl.java
@@ -100,7 +100,7 @@ final class VmClassLoaderImpl extends VmObjectImpl implements VmClassLoader {
             throw duplicateClass(vm);
         }
         MemoryImpl memory = (MemoryImpl) content.getMemory();
-        ClassFile classFile = ClassFile.of(classContext, ByteBuffer.wrap(memory.getArray(), content.getArrayElementOffset(0), content.getLength()));
+        ClassFile classFile = ClassFile.of(classContext, ByteBuffer.wrap(memory.getArray(), Math.toIntExact(content.getArrayElementOffset(0)), content.getLength()));
         // todo: proper verification...
         DefinedTypeDefinition.Builder builder = classContext.newTypeBuilder();
         classFile.accept(builder);
